users tests

+++ user/views/vuser.py
@@ -18,7 +18,7 @@ from django.contrib.auth.models import User
 # local Django
 from user.serializers import UserSerializer, UserRegisterSerializer, UserHeavySerializer
 
-class VUserList(APIView):
+class UserListView(APIView):
     """
     ...
     """
@@ -55,7 +55,7 @@ class VUserList(APIView):
         else:
             return Response(response.errors, status=status.HTTP_400_BAD_REQUEST)
 
-class VUserDetail(APIView):
+class UserDetailView(APIView):
     """
     ...
     """
@@ -97,6 +97,15 @@ class VUserDetail(APIView):
         ...
         """
         user = self.get_object(pk)
-        # value = user.id
+
+        if user.id == request.user.id:
+            return Response("can't delete himself", status=status.HTTP_400_BAD_REQUEST)
+
+        if user.is_superuser == True:
+            return Response('super users cannot be deleted', status=status.HTTP_400_BAD_REQUEST)
+
+        if user.is_staff == True and request.user.is_superuser != True:
+            return Response('user cannot delete administrators', status=status.HTTP_400_BAD_REQUEST)
+
         user.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
